tracktrains/requirements/base.txt
Django==1.6.5
South==0.8.4
django-tastypie==0.11.1
